Score Tracker Update
Implemented a counter for the score.
Added difficulty and streak trackers to the Map Area.
Added logic to increment/decrement streak, difficulty, and score on answers.
Added placeholder logic to select question based on difficulty.
Added hud controller and object.
Score and coins are now managed by the hud controller.