Encounter animation added
Added a flashing sprite on top of the player when a battle begins.

Currently locks the player's movement before the sprite shows

Currently using a sprite I found in the folder, need to add a exclamation mark sprite to use.

I think a flashing sprite is better at showing the player has encountered something rather than a static sprite and then the battle begins